+++ apps/cowswap-frontend/src/common/updaters/CancelReplaceTxUpdater.tsx
@@ -72,6 +72,8 @@ function unwatchTxChanges(_pendingHashes: string[], _chainId: number) {
  * TODO: find a new provider https://github.com/cowprotocol/cowswap/issues/4901
  */
 export function CancelReplaceTxUpdater(): null {
+  // TODO M-6 COW-573
+  // This flow will be reviewed and updated later, to include a wagmi alternative
   const provider = useWalletProvider()
   const { chainId, account } = useWalletInfo()
   const dispatch = useAppDispatch()