+++ apps/cowswap-frontend/src/cowSdk.ts
@@ -1,21 +1,31 @@
 import { useEffect } from 'react'
 
-import { getRpcProvider } from '@cowprotocol/common-const'
+import { getRpcProvider, LAUNCH_DARKLY_VIEM_MIGRATION } from '@cowprotocol/common-const'
 import { getCurrentChainIdFromUrl, isBarnBackendEnv } from '@cowprotocol/common-utils'
-import { DEFAULT_BACKOFF_OPTIONS, MetadataApi, OrderBookApi, setGlobalAdapter } from '@cowprotocol/cow-sdk'
+import {
+  DEFAULT_BACKOFF_OPTIONS,
+  MetadataApi,
+  OrderBookApi,
+  setGlobalAdapter,
+  AbstractProviderAdapter,
+} from '@cowprotocol/cow-sdk'
+import { PERMIT_ACCOUNT } from '@cowprotocol/permit-utils'
 import { EthersV5Adapter } from '@cowprotocol/sdk-ethers-v5-adapter'
+import { ViemAdapter } from '@cowprotocol/sdk-viem-adapter'
 import { useWeb3React } from '@web3-react/core'
+
+import { usePublicClient, useWalletClient } from 'wagmi'
 
 const chainId = getCurrentChainIdFromUrl()
 const prodBaseUrls = process.env.REACT_APP_ORDER_BOOK_URLS
   ? JSON.parse(process.env.REACT_APP_ORDER_BOOK_URLS)
   : undefined
 
-export const adapter = new EthersV5Adapter({
+const legacyAdapter = new EthersV5Adapter({
   provider: getRpcProvider(chainId)!,
 })
 
-setGlobalAdapter(adapter)
+setGlobalAdapter(legacyAdapter)
 
 export const orderBookApi = new OrderBookApi({
   env: isBarnBackendEnv ? 'staging' : 'prod',
@@ -26,12 +36,26 @@ export const orderBookApi = new OrderBookApi({
 export const metadataApiSDK = new MetadataApi()
 
 export function CowSdkUpdater(): null {
+  const publicClient = usePublicClient()
+  const { data: walletClient } = useWalletClient()
   const { chainId, provider, account } = useWeb3React()
 
   useEffect(() => {
+    if (!LAUNCH_DARKLY_VIEM_MIGRATION) return
+    if (!publicClient) return
+    if (walletClient) {
+      // TODO: fix the type casting
+      setGlobalAdapter(new ViemAdapter({ provider: publicClient, walletClient }) as AbstractProviderAdapter)
+    } else {
+      setGlobalAdapter(new ViemAdapter({ provider: publicClient, signer: PERMIT_ACCOUNT }) as AbstractProviderAdapter)
+    }
+  }, [publicClient, walletClient])
+
+  useEffect(() => {
+    if (LAUNCH_DARKLY_VIEM_MIGRATION) return
     if (!provider) return
-    adapter.setProvider(provider)
-    adapter.setSigner(provider.getSigner())
+    legacyAdapter.setProvider(provider)
+    legacyAdapter.setSigner(provider.getSigner())
   }, [chainId, account, provider])
 
   return null