+++ apps/cowswap-frontend/src/modules/limitOrders/hooks/useTradeFlowContext.ts
@@ -24,10 +24,11 @@ import { useSafeMemo } from 'common/hooks/useSafeMemo'
 
 import { useLimitOrdersDerivedState } from './useLimitOrdersDerivedState'
 
-
 // TODO: Break down this large function into smaller functions
 // eslint-disable-next-line max-lines-per-function
 export function useTradeFlowContext(): TradeFlowContext | null {
+  // TODO M-6 COW-573
+  // This flow will be reviewed and updated later, to include a wagmi alternative
   const provider = useWalletProvider()
   const { account } = useWalletInfo()
   const { allowsOffchainSigning } = useWalletDetails()