+++ apps/cowswap-frontend/src/modules/permit/hooks/useGeneratePermitHook.ts
@@ -34,6 +34,8 @@ export function useGeneratePermitHook(): GeneratePermitHook {
   useAtomValue(staticPermitCacheAtom)
   useAtomValue(userPermitCacheAtom)
 
+  // TODO M-6 COW-573
+  // This flow will be reviewed and updated later, to include a wagmi alternative
   const provider = useWalletProvider()
 
   return useCallback(